Reorganize code to better structure in packages
