+++ lib/main.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:patta/api/api.dart' as api;
-import 'package:patta/data_model/inspiration_card.dart';
-import 'package:patta/inspiration_card.dart';
 import 'package:patta/local_database/database.dart';
 import 'package:patta/resources/strings.dart';
+import 'package:patta/ui/screens/today/TodayScreen.dart';
 import 'package:provider/provider.dart';
 
 void main() => runApp(MyApp());
@@ -22,82 +20,7 @@ class MyApp extends StatelessWidget {
           primaryColor: Color(0xffdcd3c0),
           accentColor: Color(0xff6d695f),
         ),
-        home: MyHomePage(title: strings['en'].appName),
-      ),
-    );
-  }
-}
-
-class MyHomePage extends StatelessWidget {
-  final String title;
-
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Color(0xfff4efe7),
-      appBar: AppBar(
-        title: Text(
-          title,
-          style: TextStyle(
-            inherit: true,
-            color: Color(0xff6d695f),
-          ),
-        ),
-      ),
-      body: FutureBuilder<List<CardModel>>(
-        future: api.fetchToday(),
-        builder: (
-          BuildContext context,
-          AsyncSnapshot<List<CardModel>> snapshot,
-        ) {
-          if (snapshot.hasData) {
-            return ListView(
-              children: snapshot.data
-                  .map((card) {
-                    if (card is InspirationCardModel) {
-                      return InspirationCard(
-                        card,
-                        Provider.of<PariyattiDatabase>(context),
-                      );
-                    } else {
-                      return null;
-                    }
-                  })
-                  .where((widget) => (widget != null))
-                  .toList(),
-            );
-          } else if (snapshot.hasError) {
-            return Center(
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: Icon(
-                      Icons.error,
-                      color: Color(0xff6d695f),
-                    ),
-                  ),
-                  Text(
-                    strings['en'].errorMessageTryAgainLater,
-                    style: TextStyle(
-                      inherit: true,
-                      color: Color(0xff6d695f),
-                      fontSize: 16.0,
-                    ),
-                  )
-                ],
-              ),
-            );
-          } else {
-            return Center(
-              child: CircularProgressIndicator(),
-            );
-          }
-        },
+        home: TodayScreen(title: strings['en'].appName),
       ),
     );
   }
