+++ lib/ui/model/CardModel.dart
@@ -1,0 +1,5 @@
+abstract class CardModel {
+  final String id;
+
+  CardModel(this.id);
+}
