+++ lib/ui/screens/today/cards/InspirationCard.dart
@@ -1,10 +1,10 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
-import 'package:patta/common_widgets/bookmark_button.dart';
-import 'package:patta/data_model/inspiration_card.dart';
 import 'package:patta/local_database/database.dart';
 import 'package:patta/resources/strings.dart';
+import 'package:patta/ui/common_widgets/bookmark_button.dart';
+import 'package:patta/ui/model/InspirationCardModel.dart';
 import 'package:patta/util.dart';
 import 'package:wc_flutter_share/wc_flutter_share.dart';
